LayoutTests/printing/page-rule-selection-expected.txt

This tests page style selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default margin is 1in = 96px.
PASS layoutTestController.pageProperty('margin-left', 0) is "96"
PASS layoutTestController.pageProperty('margin-left', 1) is "96"
PASS layoutTestController.pageProperty('margin-left', 2) is "96"
PASS layoutTestController.pageProperty('margin-left', 3) is "96"
PASS layoutTestController.pageProperty('margin-left', 4) is "96"
Set margin to 200px via user style sheet.
PASS layoutTestController.pageProperty('margin-left', 0) is "200"
PASS layoutTestController.pageProperty('margin-left', 1) is "200"
PASS layoutTestController.pageProperty('margin-left', 2) is "200"
PASS layoutTestController.pageProperty('margin-left', 3) is "200"
PASS layoutTestController.pageProperty('margin-left', 4) is "200"
Set margin to 100px.
PASS layoutTestController.pageProperty('margin-left', 0) is "100"
PASS layoutTestController.pageProperty('margin-left', 1) is "100"
PASS layoutTestController.pageProperty('margin-left', 2) is "100"
PASS layoutTestController.pageProperty('margin-left', 3) is "100"
PASS layoutTestController.pageProperty('margin-left', 4) is "100"
Set margin for right pages to 50px. The first page is a right page.
PASS layoutTestController.pageProperty('margin-left', 0) is "50"
PASS layoutTestController.pageProperty('margin-left', 1) is "100"
PASS layoutTestController.pageProperty('margin-left', 2) is "50"
PASS layoutTestController.pageProperty('margin-left', 3) is "100"
PASS layoutTestController.pageProperty('margin-left', 4) is "50"
Set margin for left pages to 75px.
PASS layoutTestController.pageProperty('margin-left', 0) is "50"
PASS layoutTestController.pageProperty('margin-left', 1) is "75"
PASS layoutTestController.pageProperty('margin-left', 2) is "50"
PASS layoutTestController.pageProperty('margin-left', 3) is "75"
PASS layoutTestController.pageProperty('margin-left', 4) is "50"
Set margin for the first page to 125px.
PASS layoutTestController.pageProperty('margin-left', 0) is "125"
PASS layoutTestController.pageProperty('margin-left', 1) is "75"
PASS layoutTestController.pageProperty('margin-left', 2) is "50"
PASS layoutTestController.pageProperty('margin-left', 3) is "75"
PASS layoutTestController.pageProperty('margin-left', 4) is "50"
Change the document direction to RTL. This makes the first page a left page.
PASS layoutTestController.pageProperty('margin-left', 0) is "125"
PASS layoutTestController.pageProperty('margin-left', 1) is "50"
PASS layoutTestController.pageProperty('margin-left', 2) is "75"
PASS layoutTestController.pageProperty('margin-left', 3) is "50"
PASS layoutTestController.pageProperty('margin-left', 4) is "75"
Check font update and line height override.
PASS layoutTestController.pageProperty('line-height', 0) is "180"
PASS layoutTestController.pageProperty('font-family', 0) is "Arial"
PASS layoutTestController.pageProperty('font-size', 0) is "30"
PASS successfullyParsed is true

TEST COMPLETE
